Criação do contexto relacionado com a bateria; ALteração de algumas classes;

+++ app/src/main/java/pt/ulusofona/ecati/deisi/licenciatura/cm2021/grupo2/ui/listeners/OnBatteryCurrentListener.kt
@@ -2,5 +2,5 @@ package pt.ulusofona.ecati.deisi.licenciatura.cm2021.grupo2.ui.listeners
 
 interface OnBatteryCurrentListener {
 
-    fun onCurrentChanged(current: Double)
+    fun onCurrentChanged(current: Int)
 }
